+++ Sources/Utils/ModelUtils/KspLayer.cs
@@ -6,34 +6,38 @@ using System;
 
 namespace KSPDev.ModelUtils {
 
-/// <summary>Flags for the collision layers in KSP.</summary>
+/// <summary>Defines various layers used in the game.</summary>
 /// <remarks>
 /// It's not a full set of the layers. More investigation is needed to reveal all of them.
 /// </remarks>
-[Flags]
-public enum KspLayerMask {
-  /// <summary>Just a default value that doesn't match any layer.</summary>
-  NONE = 0,
-  
-  /// <summary>Layer for a regular part.</summary>
-  PARTS = 1 << 0,
-  
-  /// <summary>Layer to set bounds of a celestial body.</summary>
+/// <seealso cref="KspLayerMask"/>
+public enum KspLayer {
+  /// <summary>The layer for a regular part.</summary>
+  Part = 0,
+
+  /// <summary>The layer to set bounds of a celestial body.</summary>
   /// <remarks>
-  /// A very rough boundary of a planet, moon or asteroid. Used for macro objects detection.
+  /// It's a very rough boundary of a planet, moon or asteroid. Used for macro objects detection.
   /// </remarks>
-  SERVICE_LAYER = 1 << 10, 
+  Service = 10, 
 
-  /// <summary>"Zero" level of a static structure on the surface.</summary>
+  /// <summary>The "zero" level collider of a static structure on the surface.</summary>
   /// <remarks>E.g. a launchpad.</remarks>
-  SURFACE = 1 << 15,
+  SurfaceCollider = 15,
 
-  /// <summary>Layer of kerbonaut models.</summary>
-  KERBALS = 1 << 17,
+  /// <summary>The layer for the kerbonaut models.</summary>
+  Kerbal = 17,
 
-  /// <summary>A layer for FX.</summary>
+  /// <summary>The layer for the various interaction colliders.</summary>
+  /// <remarks>
+  /// The meshes on this layer are not rendered. They are only used to trigger collision events when
+  /// a kerbal model is in range.
+  /// </remarks>
+  TriggerCollider = 21,
+
+  /// <summary>The layer for FX objects.</summary>
   /// <remarks>E.g. <c>PadFXReceiver</c> on the Kerbins VAB launchpad.</remarks>
-  FX = 1 << 30,
-};
+  Fx = 30,
+}
 
 }  // namespace