+++ Sources/Utils/ModelUtils/Meshes.cs
@@ -83,7 +83,7 @@ public static class Meshes {
   /// <seealso href="https://docs.unity3d.com/ScriptReference/Transform.html">Unity3D: Transform
   /// </seealso>
   public static GameObject CreateCylinder(
-      float diameter, float length, Material material,Transform parent,
+      float diameter, float length, Material material, Transform parent,
       Colliders.PrimitiveCollider colliderType = Colliders.PrimitiveCollider.None) {
     // Default length scale is 2.0.
     var obj = Meshes.CreatePrimitive(PrimitiveType.Cylinder,